refactor: Input 컴포넌트에서 style코드 분리
