+++ src/components/atoms/Input/Input.tsx
@@ -1,24 +1,5 @@
 import { InputHTMLAttributes } from "react";
-import styled from "styled-components";
-
-const SInput = styled.input<{ isError: boolean }>`
-  width: 100%;
-  height: 15px;
-  padding: 8px 15px 8px 8px;
-  gap: 8px;
-  border: 1px solid
-    ${(props) => (props.isError ? "red" : "rgba(172, 172, 172, 1)")};
-  border-radius: 2px;
-  font-family: Inter;
-  font-size: 11px;
-  font-weight: 400;
-  line-height: 14.88px;
-  text-align: left;
-
-  &::placeholder {
-    color: rgba(172, 172, 172, 1);
-  }
-`;
+import * as S from "./style";
 
 interface Props extends InputHTMLAttributes<HTMLInputElement> {
   isError: boolean;
@@ -36,7 +17,7 @@ export default function Input({
   onBlurInput,
 }: Props) {
   return (
-    <SInput
+    <S.Input
       type={type ? type : "text"}
       maxLength={maxLength}
       placeholder={placeholder}
@@ -48,6 +29,6 @@ export default function Input({
       onBlur={(e) => {
         if (onBlurInput) onBlurInput(e.target.value);
       }}
-    ></SInput>
+    ></S.Input>
   );
 }
