+++ src/components/atoms/Input/style.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const Input = styled.input<{ isError: boolean }>`
+  width: 100%;
+  height: 15px;
+  padding: 8px 15px 8px 8px;
+  gap: 8px;
+  border: 1px solid
+    ${(props) => (props.isError ? "red" : "rgba(172, 172, 172, 1)")};
+  border-radius: 2px;
+  font-family: Inter;
+  font-size: 11px;
+  font-weight: 400;
+  line-height: 14.88px;
+  text-align: left;
+
+  &::placeholder {
+    color: rgba(172, 172, 172, 1);
+  }
+`;
